Levanta el servidor

+++ tpoDatos/src/main/repository/UserRepository.java
@@ -3,7 +3,7 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends MongoRepository<User, Integer>{
+public interface UserRepository extends MongoRepository<User,Long>{
 
     public User findByUsername(String username);
     public User findByEmail(String email);
